data-pipeline-project/requirements.txt
pandas
google-cloud-bigquery
boto3
sqlalchemy
pyyaml
schedule
datetime